Beauty/requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.7
django-admin-thumbnails==0.2.9
django-cleanup==8.1.0
flake8==7.1.0
flake8-bugbear==24.4.26
mccabe==0.7.0
pep8-naming==0.14.1
pillow==10.4.0
pycodestyle==2.12.0
pyflakes==3.2.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
Unidecode==1.3.8
